changes to find_orf.py
added some changes. worked at one point but now I have no idea what is
going on. cant get it to work but so over it that I am not gonna work
on this anymore. not quite sure the problem since the error is in an
indent, which i have changed and edited and re-written at least 100 times.
I'm done.